requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.30
python3-nmap
python-nmap
flask_cors
netifaces
uvicorn==0.35.0
zeroconf==0.147.0
